requirements.txt
Flask==3.1.2
mysql-connector-python==9.4.0
python-dotenv==1.1.1
twilio==9.8.4
flask-cors==6.0.1
razorpay==2.0.0

